refactor: split remaining test modules
