+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.34.3] - 2026-06-27
+
 ## [1.34.2] - 2026-06-25
 
 ### Fixed
