+++ Cargo.toml
@@ -4,7 +4,7 @@ resolver = "2"
 
 [package]
 name = "cortex"
-version = "1.34.2"
+version = "1.34.3"
 edition = "2024"
 rust-version = "1.86"
 license = "MIT"
