+++ server.json
@@ -7,11 +7,11 @@
     "url": "https://github.com/jmagar/cortex",
     "source": "github"
   },
-  "version": "1.34.2",
+  "version": "1.34.3",
   "packages": [
     {
       "registryType": "oci",
-      "identifier": "ghcr.io/jmagar/cortex:v1.34.2",
+      "identifier": "ghcr.io/jmagar/cortex:v1.34.3",
       "transport": {
         "type": "stdio"
       },
