+++ src/assessment.rs
@@ -375,144 +375,8 @@ async fn read_bounded_stderr(stderr: tokio::process::ChildStderr) -> std::io::Re
     }
 }
 
-#[derive(Default)]
-pub(crate) struct GeminiStreamState {
-    text: String,
-    result_text: Option<String>,
-    saw_success: bool,
-}
-
-impl GeminiStreamState {
-    pub(crate) fn handle_line<F>(&mut self, line: &str, on_delta: &mut F) -> Result<()>
-    where
-        F: FnMut(&str) -> Result<()> + Send,
-    {
-        let trimmed = line.trim();
-        if trimmed.is_empty() {
-            return Ok(());
-        }
-        let value: Value = serde_json::from_str(trimmed)
-            .map_err(|err| anyhow!("malformed Gemini stream JSON: {err}: {trimmed}"))?;
-        match value.get("type").and_then(Value::as_str) {
-            Some("tool_use") => self.handle_tool_use(&value)?,
-            Some("tool_result") => {}
-            Some("error") => bail!("Gemini headless stream error: {value}"),
-            Some("message") if value.get("role").and_then(Value::as_str) == Some("assistant") => {
-                if let Some(delta) = message_content(&value) {
-                    self.push_delta(&delta, on_delta)?;
-                }
-            }
-            Some("result") => self.handle_result(&value)?,
-            _ if contains_tool_event(&value) => {
-                bail!("Gemini headless emitted a tool event in assessment-only mode");
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-
-    fn handle_result(&mut self, value: &Value) -> Result<()> {
-        if value.get("status").and_then(Value::as_str) != Some("success") {
-            bail!("Gemini headless returned unsuccessful result: {value}");
-        }
-        if let Some(text) = value.get("response").and_then(Value::as_str) {
-            if !text.trim().is_empty() {
-                self.result_text = Some(text.to_string());
-            }
-        }
-        self.saw_success = true;
-        Ok(())
-    }
-
-    fn handle_tool_use(&mut self, value: &Value) -> Result<()> {
-        let tool_name = value
-            .get("name")
-            .and_then(Value::as_str)
-            .or_else(|| value.get("tool_name").and_then(Value::as_str));
-        match tool_name {
-            Some("activate_skill") | Some("update_topic") => Ok(()),
-            Some("write_file") => {
-                let Some(content) = value
-                    .get("parameters")
-                    .and_then(|params| params.get("content"))
-                    .and_then(Value::as_str)
-                    .filter(|content| !content.trim().is_empty())
-                else {
-                    bail!(
-                        "Gemini headless emitted write_file without assessment content; raw event: {value}"
-                    );
-                };
-                self.result_text = Some(content.to_string());
-                Ok(())
-            }
-            _ => bail!(
-                "Gemini headless emitted unexpected tool call '{}' in assessment mode; raw event: {value}",
-                tool_name.unwrap_or("unknown")
-            ),
-        }
-    }
-
-    fn push_delta<F>(&mut self, delta: &str, on_delta: &mut F) -> Result<()>
-    where
-        F: FnMut(&str) -> Result<()> + Send,
-    {
-        if delta.is_empty() {
-            return Ok(());
-        }
-        self.text.push_str(delta);
-        on_delta(delta)
-    }
-
-    pub(crate) fn finish(self) -> Result<String> {
-        if !self.saw_success {
-            bail!("Gemini headless stream ended without a success result");
-        }
-        if let Some(result) = self.result_text {
-            if !result.trim().is_empty() {
-                return Ok(result);
-            }
-        }
-        if !self.text.trim().is_empty() {
-            return Ok(self.text);
-        }
-        bail!("Gemini headless returned no assessment text");
-    }
-}
-
-fn message_content(value: &Value) -> Option<String> {
-    if let Some(content) = value.get("content").and_then(Value::as_str) {
-        return Some(content.to_string());
-    }
-    if let Some(parts) = value.get("content").and_then(Value::as_array) {
-        let mut out = String::new();
-        for part in parts {
-            if let Some(text) = part.as_str() {
-                out.push_str(text);
-            } else if let Some(text) = part.get("text").and_then(Value::as_str) {
-                out.push_str(text);
-            }
-        }
-        return (!out.is_empty()).then_some(out);
-    }
-    None
-}
-
-fn contains_tool_event(value: &Value) -> bool {
-    match value {
-        Value::String(s) => matches!(s.as_str(), "tool_use" | "tool_result"),
-        Value::Array(items) => items.iter().any(contains_tool_event),
-        Value::Object(map) => map.iter().any(|(key, value)| {
-            key == "tool_use"
-                || key == "tool_result"
-                || (key == "type"
-                    && value
-                        .as_str()
-                        .is_some_and(|s| s == "tool_use" || s == "tool_result"))
-                || contains_tool_event(value)
-        }),
-        _ => false,
-    }
-}
+mod gemini_stream;
+pub(crate) use gemini_stream::GeminiStreamState;
 
 fn env_or_default(var_name: &str, default_program: &str) -> String {
     non_empty_env(var_name).unwrap_or_else(|| default_program.to_string())
@@ -563,298 +427,5 @@ fn looks_secretish(token: &str) -> bool {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use serial_test::serial;
-
-    struct EnvGuard {
-        name: &'static str,
-        old: Option<std::ffi::OsString>,
-    }
-
-    impl EnvGuard {
-        fn set(name: &'static str, value: impl AsRef<std::ffi::OsStr>) -> Self {
-            let old = std::env::var_os(name);
-            // TODO: Audit that the environment access only happens in single-threaded code.
-            unsafe { std::env::set_var(name, value) };
-            Self { name, old }
-        }
-    }
-
-    impl Drop for EnvGuard {
-        fn drop(&mut self) {
-            match &self.old {
-                // TODO: Audit that the environment access only happens in single-threaded code.
-                Some(value) => unsafe { std::env::set_var(self.name, value) },
-                // TODO: Audit that the environment access only happens in single-threaded code.
-                None => unsafe { std::env::remove_var(self.name) },
-            }
-        }
-    }
-
-    #[test]
-    fn embedded_assessment_skill_has_injection_defense() {
-        assert!(!SKILL_MD.trim().is_empty());
-        assert!(SKILL_MD.contains("untrusted input"));
-        assert!(SKILL_MD.contains("Never attribute blame without citing specific evidence"));
-        assert!(SKILL_MD.contains("Real frustration with incidental profanity"));
-        assert!(SKILL_MD.contains("Trend evidence unavailable"));
-        assert!(SKILL_MD
-            .contains("Never infer recurrence, isolation, or system-wide behavior from a single incident without comparison evidence"));
-        assert!(SKILL_MD.contains("Never combine \"Trend evidence unavailable\""));
-        assert!(SKILL_MD.contains("write exactly: **Trend evidence unavailable.**"));
-        assert!(SKILL_MD.contains("Never write \"no systemic failure\""));
-        assert!(SKILL_MD.contains("Never collapse \"real frustration with incidental profanity\""));
-        assert!(
-            SKILL_MD.contains("The executive summary must preserve the same uncertainty level")
-        );
-    }
-
-    #[test]
-    fn assessment_prompt_references_skill_and_wraps_evidence() {
-        let prompt = build_assessment_prompt(r#"{"incident_id":"inc-1"}"#);
-        assert!(prompt.contains("cortex-frustration-assessment"));
-        assert!(prompt.contains("Do not write files"));
-        assert!(prompt.contains("Use this skill after running"));
-        assert!(prompt.contains("<untrusted-evidence"));
-        assert!(prompt.contains(r#""incident_id":"inc-1""#));
-    }
-
-    #[test]
-    fn gemini_command_spec_uses_stdin_stream_json() {
-        let config = GeminiAssessConfig {
-            program: "gemini".into(),
-            model: "gemini-test".into(),
-            source_home: None,
-            timeout_secs: 5,
-        };
-        let spec = config.command_spec().unwrap();
-        assert_eq!(spec.program, "gemini");
-        assert_eq!(spec.output_mode, "stream-json");
-        assert!(
-            spec.args
-                .windows(2)
-                .any(|w| w == ["--prompt", GEMINI_STDIN_PROMPT_STUB])
-        );
-        assert!(
-            spec.args
-                .windows(2)
-                .any(|w| w == ["--approval-mode", "plan"])
-        );
-        assert!(spec.args.windows(2).any(|w| w == ["--extensions", "none"]));
-        assert!(
-            spec.args
-                .windows(2)
-                .any(|w| w == ["--output-format", "stream-json"])
-        );
-        assert!(
-            spec.args
-                .windows(2)
-                .any(|w| w == ["--model", "gemini-test"])
-        );
-    }
-
-    #[test]
-    fn isolated_gemini_home_clears_side_effect_settings_and_installs_skill() {
-        let source = tempfile::tempdir().unwrap();
-        let gemini_dir = source.path().join(".gemini");
-        fs::create_dir_all(&gemini_dir).unwrap();
-        fs::write(
-            gemini_dir.join("settings.json"),
-            r#"{
-                "security": {"auth": {"selectedType": "oauth-personal"}},
-                "mcpServers": {"danger": {}},
-                "hooks": {"SessionStart": []},
-                "context": {"fileName": ["GEMINI.md"]},
-                "admin": {"enabled": true}
-            }"#,
-        )
-        .unwrap();
-        fs::write(gemini_dir.join("oauth_creds.json"), "{}").unwrap();
-
-        let config = GeminiAssessConfig {
-            program: "gemini".into(),
-            model: "gemini-test".into(),
-            source_home: Some(source.path().to_path_buf()),
-            timeout_secs: 5,
-        };
-        let home = prepare_gemini_home(&config).unwrap();
-        let isolated = home.path().join(".gemini");
-        assert!(isolated.join("oauth_creds.json").is_file());
-        assert!(
-            isolated
-                .join("skills")
-                .join(SKILL_NAME)
-                .join("SKILL.md")
-                .is_file()
-        );
-
-        let settings: Value =
-            serde_json::from_slice(&fs::read(isolated.join("settings.json")).unwrap()).unwrap();
-        assert_eq!(settings["mcpServers"], json!({}));
-        assert_eq!(settings["hooks"], json!({}));
-        assert_eq!(settings["context"], json!({ "fileName": [] }));
-        assert!(settings.get("admin").is_none());
-        assert_eq!(
-            settings["security"]["auth"]["selectedType"],
-            json!("oauth-personal")
-        );
-    }
-
-    #[test]
-    fn stream_parser_accepts_deltas_and_result_text() {
-        let mut parser = GeminiStreamState::default();
-        let mut streamed = String::new();
-        parser
-            .handle_line(
-                r#"{"type":"message","role":"assistant","content":[{"text":"Hello "},{"text":"world"}]}"#,
-                &mut |delta| {
-                    streamed.push_str(delta);
-                    Ok(())
-                },
-            )
-            .unwrap();
-        parser
-            .handle_line(r#"{"type":"result","status":"success"}"#, &mut |_| Ok(()))
-            .unwrap();
-        assert_eq!(streamed, "Hello world");
-        assert_eq!(parser.finish().unwrap(), "Hello world");
-    }
-
-    #[test]
-    fn stream_parser_rejects_unexpected_tool_call() {
-        let mut parser = GeminiStreamState::default();
-        let err = parser
-            .handle_line(
-                r#"{"type":"tool_use","tool_name":"shell","parameters":{"cmd":"date"}}"#,
-                &mut |_| Ok(()),
-            )
-            .unwrap_err()
-            .to_string();
-        assert!(err.contains("unexpected tool call"));
-    }
-
-    #[test]
-    fn stream_parser_recovers_markdown_from_write_file_tool_call() {
-        let mut parser = GeminiStreamState::default();
-        parser
-            .handle_line(
-                r##"{"type":"tool_use","tool_name":"write_file","parameters":{"file_path":"/tmp/syslog-gemini-headless-x/.gemini/tmp/session/plans/frustration_assessment.md","content":"# Frustration Assessment\n\nRecovered report."}}"##,
-                &mut |_| Ok(()),
-            )
-            .unwrap();
-        parser
-            .handle_line(r#"{"type":"result","status":"success"}"#, &mut |_| Ok(()))
-            .unwrap();
-        assert_eq!(
-            parser.finish().unwrap(),
-            "# Frustration Assessment\n\nRecovered report."
-        );
-    }
-
-    #[test]
-    fn stream_parser_prefers_write_file_assessment_over_preamble() {
-        let mut parser = GeminiStreamState::default();
-        let mut streamed = String::new();
-        parser
-            .handle_line(
-                r#"{"type":"message","role":"assistant","content":[{"text":"I will write the assessment now."}]}"#,
-                &mut |delta| {
-                    streamed.push_str(delta);
-                    Ok(())
-                },
-            )
-            .unwrap();
-        parser
-            .handle_line(
-                r##"{"type":"tool_use","tool_name":"write_file","parameters":{"file_path":"/tmp/syslog-gemini-headless-x/.gemini/tmp/session/plans/frustration_assessment.md","content":"# Frustration Assessment\n\nRecovered report."}}"##,
-                &mut |_| Ok(()),
-            )
-            .unwrap();
-        parser
-            .handle_line(r#"{"type":"result","status":"success"}"#, &mut |_| Ok(()))
-            .unwrap();
-        assert_eq!(streamed, "I will write the assessment now.");
-        assert_eq!(
-            parser.finish().unwrap(),
-            "# Frustration Assessment\n\nRecovered report."
-        );
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn gemini_assessment_timeout_kills_and_reaps_child() {
-        let source = tempfile::tempdir().unwrap();
-        fs::create_dir_all(source.path().join(".gemini")).unwrap();
-        fs::write(source.path().join(".gemini").join("settings.json"), "{}").unwrap();
-        let script = source.path().join("fake-gemini.sh");
-        fs::write(&script, "#!/usr/bin/env bash\nsleep 5\n").unwrap();
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let mut perms = fs::metadata(&script).unwrap().permissions();
-            perms.set_mode(0o755);
-            fs::set_permissions(&script, perms).unwrap();
-        }
-
-        let config = GeminiAssessConfig {
-            program: script.display().to_string(),
-            model: "gemini-test".into(),
-            source_home: Some(source.path().to_path_buf()),
-            timeout_secs: 1,
-        };
-        let err = run_gemini_assessment("prompt", &config, |_| Ok(()))
-            .await
-            .unwrap_err()
-            .to_string();
-        assert!(err.contains("timed out after 1 seconds"));
-        assert!(err.contains("cleanup:"));
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn gemini_assessment_reports_child_stderr_before_stdin_pipe_error() {
-        let source = tempfile::tempdir().unwrap();
-        fs::create_dir_all(source.path().join(".gemini")).unwrap();
-        fs::write(source.path().join(".gemini").join("settings.json"), "{}").unwrap();
-        let script = source.path().join("fake-gemini.sh");
-        fs::write(
-            &script,
-            "#!/usr/bin/env bash\necho 'simulated gemini failure' >&2\nexit 42\n",
-        )
-        .unwrap();
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let mut perms = fs::metadata(&script).unwrap().permissions();
-            perms.set_mode(0o755);
-            fs::set_permissions(&script, perms).unwrap();
-        }
-
-        let config = GeminiAssessConfig {
-            program: script.display().to_string(),
-            model: "gemini-test".into(),
-            source_home: Some(source.path().to_path_buf()),
-            timeout_secs: 5,
-        };
-        let large_prompt = "x".repeat(2 * 1024 * 1024);
-        let err = run_gemini_assessment(&large_prompt, &config, |_| Ok(()))
-            .await
-            .unwrap_err()
-            .to_string();
-        assert!(err.contains("Gemini headless exited"));
-        assert!(err.contains("simulated gemini failure"));
-    }
-
-    #[test]
-    #[serial]
-    fn env_config_uses_syslog_specific_knobs() {
-        let _cmd = EnvGuard::set("CORTEX_HEADLESS_GEMINI_CMD", "custom-gemini");
-        let _model = EnvGuard::set("CORTEX_HEADLESS_GEMINI_MODEL", "gemini-custom");
-        let _timeout = EnvGuard::set("CORTEX_LLM_COMPLETION_TIMEOUT_SECS", "7");
-        let config = GeminiAssessConfig::from_env(None);
-        assert_eq!(config.program, "custom-gemini");
-        assert_eq!(config.model, "gemini-custom");
-        assert_eq!(config.timeout_secs, 7);
-    }
-}
+#[path = "assessment_tests.rs"]
+mod tests;
